backend/pyproject.toml
[tool.poetry]
name = "basic-receipt-ocr"
version = "0.1.0"
description = "An OCR for extracting text from receipts for analysis"
authors = ["smtaha512 <s.m.taha10@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-vision = "^3.7.4"
ruff = "^0.7.1"
fastapi = {extras = ["standard"], version = "^0.115.3"}
python-multipart = "^0.0.12"
fuzzywuzzy = "^0.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
